Transfer bar list fragment, update naming convetion
